--- archive/2d-vectors/2.1/vectors.py
@@ -1,12 +1,0 @@
-from vector_drawing import *
-dino_vectors = [
-    (6, 4), (3, 1), (1, 2), (-1, 5), (-2, 5), (-3, 4), (-4, 4),
-    (-5, 3), (-5, 2), (-2, 2), (-5, 1), (-4, 0), (-2, 1), (-1, 0), (0, -3),
-    (-1, -4), (1, -4), (2, -3), (1, -2), (3, -1), (5, 1)
-]
-
-draw(
-    Points(*[(x, x**2) for x in range(-10, 11)]),
-    grid=(1, 10),
-    nice_aspect_ratio=False,
-)
