--- archive/2d-vectors/README.md
@@ -1,16 +1,0 @@
-# Important
-Exercise solution implementations can be found in the ```vectors_ex.py``` files. All answers can be found in the docx document as well.
-Some of the new Python updates have broken the titles of the windows.
-
-# Running
-Please install the Anaconda Python distribution when running the Python files in this repository (choose Python 3.7): https://www.anaconda.com/products/distribution
-
-# Key Terms
-- 2D Space: *plane*
-    - height
-    - width
-
-A refernce point (**origin**) is required to describe a point in a 2d plane.
-
-- Two-Dimensional Vector
-Point in the plane relative to origin.
